pyproject.toml
[tool.poetry]
name = "wav2vec2_finetune_flash"
version = "0.1.0"
description = ""
authors = ["dmitriy.shyrmakin <dmitriy.shyrmakin@skillbox.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
loguru = "*"
rope = "*"
toml = "*"
pre-commit = "^2.13.0"
torch = "^1.10.1"
transformers = "^4.15.0"
torchaudio = "^0.10.0"
soundfile = "*"
torchvision = "^0.11.2"
ffmpeg = "^1.4"
llvmlite = {version = "*", python = ">=3.9,<3.10"}
numba = {version = "*", python = ">=3.9,<3.10"}
audiomentations = {version = "^0.20.0", extras = ["pydub"]}
noisereduce = "^2.0.0"
lightning-flash = "^0.6.0"
datasets = "^1.17.0"
SoundFile = "^0.10.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


